+++ backend/src/main/java/ch/cevi/db/adapter/domain/CeviEvent.java
@@ -4,7 +4,8 @@ import java.time.LocalDateTime;
 import java.util.Objects;
 
 public record CeviEvent(String id, String name, String description, String applicationLink, LocalDateTime startsAt,
-                        LocalDateTime finishAt, String group, String location, String kind, CeviEventType eventType) {
+                        LocalDateTime finishAt, String group, String location, String kind, CeviEventType eventType,
+                        int participantsCount, Integer maximumParticipants) {
     public CeviEvent {
         Objects.requireNonNull(id);
         Objects.requireNonNull(name, () -> "Name mustn't be null for event " + id);
@@ -13,5 +14,15 @@ public record CeviEvent(String id, String name, String description, String appli
         Objects.requireNonNull(startsAt, () -> "Start mustn't be null for event " + id);
         Objects.requireNonNull(group, () -> "Group mustn't be null for event " + id);
         Objects.requireNonNull(location, () -> "Location mustn't be null for event " + id);
+        if (participantsCount < 0) {
+            throw new IllegalArgumentException("The participantCount must be at least 0");
+        }
+        if (maximumParticipants != null && maximumParticipants < 1) {
+            throw new IllegalArgumentException("The maximum participants must be at least 1 if set");
+        }
+    }
+
+    public boolean hasLimitedCapacity() {
+        return maximumParticipants != null;
     }
 }