+++ backend/src/main/java/ch/cevi/db/adapter/hitobito/HitobitoEvent.java
@@ -2,7 +2,8 @@ package ch.cevi.db.adapter.hitobito;
 
 import java.util.Objects;
 
-public record HitobitoEvent(String id, String name, String description, String external_application_link, HitobitoLinks links) {
+public record HitobitoEvent(String id, String name, String description, String external_application_link, HitobitoLinks links,
+                            int participant_count, Integer maximum_participants) {
     public HitobitoEvent {
         Objects.requireNonNull(id);
         Objects.requireNonNull(name);